--- extensions/protocols/dcp/dcp-identityhub/dcp-identityhub-transform-lib/src/main/java/org/eclipse/edc/identityhub/protocols/dcp/issuer/transform/from/JsonObjectFromCredentialMessageTransformer.java
@@ -1,58 +1,0 @@
-/*
- *  Copyright (c) 2025 Cofinity-X
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Cofinity-X - initial API and implementation
- *
- */
-
-package org.eclipse.edc.identityhub.protocols.dcp.issuer.transform.from;
-
-import jakarta.json.Json;
-import jakarta.json.JsonArray;
-import jakarta.json.JsonObject;
-import org.eclipse.edc.identityhub.protocols.dcp.spi.model.CredentialMessage;
-import org.eclipse.edc.jsonld.spi.JsonLdKeywords;
-import org.eclipse.edc.jsonld.spi.JsonLdNamespace;
-import org.eclipse.edc.jsonld.spi.transformer.AbstractNamespaceAwareJsonLdTransformer;
-import org.eclipse.edc.spi.types.TypeManager;
-import org.eclipse.edc.transform.spi.TransformerContext;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
-
-public class JsonObjectFromCredentialMessageTransformer extends AbstractNamespaceAwareJsonLdTransformer<CredentialMessage, JsonObject> {
-
-    private final TypeManager typeManager;
-    private final String typeContext;
-
-    public JsonObjectFromCredentialMessageTransformer(TypeManager typeManager, String typeContext, JsonLdNamespace namespace) {
-        super(CredentialMessage.class, JsonObject.class, namespace);
-        this.typeManager = typeManager;
-        this.typeContext = typeContext;
-    }
-
-    @Override
-    public @Nullable JsonObject transform(@NotNull CredentialMessage credentialRequestMessage, @NotNull TransformerContext transformerContext) {
-
-        var credentials = typeManager.getMapper(typeContext).convertValue(credentialRequestMessage.getCredentials(), JsonArray.class);
-
-        var jsonCredentials = Json.createArrayBuilder().add(Json.createObjectBuilder()
-                        .add(JsonLdKeywords.TYPE, JsonLdKeywords.JSON)
-                        .add(JsonLdKeywords.VALUE, credentials))
-                .build();
-
-        return Json.createObjectBuilder()
-                .add(TYPE, forNamespace(CredentialMessage.CREDENTIAL_MESSAGE_TERM))
-                .add(forNamespace(CredentialMessage.CREDENTIALS_TERM), jsonCredentials)
-                .build();
-    }
-
-}
