+++ extensions/protocols/dcp/dcp-issuer/dcp-issuer-api/build.gradle.kts
@@ -32,7 +32,7 @@ dependencies {
     implementation(libs.edc.lib.transform)
     implementation(libs.edc.dcp.transform)
     implementation(libs.edc.lib.validator)
-    implementation(project(":extensions:protocols:dcp:dcp-issuer:dcp-issuer-transform-lib"))
+    implementation(project(":extensions:protocols:dcp:dcp-transform-lib"))
     implementation(libs.jakarta.rsApi)
     testImplementation(libs.edc.junit)
     testImplementation(libs.edc.jsonld)
