+++ settings.gradle.kts
@@ -71,13 +71,12 @@ include(":extensions:sts:sts-account-service-remote")
 // DCP protocol modules
 include(":extensions:protocols:dcp:dcp-spi")
 include(":extensions:protocols:dcp:dcp-issuer:dcp-issuer-spi")
-include(":extensions:protocols:dcp:dcp-issuer:dcp-issuer-transform-lib")
+include(":extensions:protocols:dcp:dcp-transform-lib")
 include(":extensions:protocols:dcp:dcp-issuer:dcp-issuer-api")
 include(":extensions:protocols:dcp:dcp-issuer:dcp-issuer-core")
 
 include(":extensions:protocols:dcp:dcp-identityhub:presentation-api")
 include(":extensions:protocols:dcp:dcp-identityhub:storage-api")
-include(":extensions:protocols:dcp:dcp-identityhub:dcp-identityhub-transform-lib")
 
 
 // Identity APIs
